Deferring RESTful calls for 50% faster bootup time
